requirements.txt
fastapi
uvicorn[standard]
pydantic
sqlalchemy
groq
requests
python-dotenv
alembic
psycopg2-binary
